components reorganization. i18next modif
