+++ client/src/components/friends/Friend.jsx
@@ -2,13 +2,13 @@
 
 import { useState, useContext, useEffect } from 'react';
 import axios from 'axios';
-import '../App.css';
-import './css/clear.css'
-import './css/dark.css'
-import Header from './parts/Header';
-import PostThumbnail_onProfile from './parts/PostThumbnail_onProfile';
-import { DarkMode } from '../context/DarkMode';
-import { User } from '../context/User';
+import '../../App.css';
+import '../css/clear.css'
+import '../css/dark.css'
+import Header from '../parts/Header';
+import PostThumbnail_onProfile from '../parts/PostThumbnail_onProfile';
+import { DarkMode } from '../../context/DarkMode';
+import { User } from '../../context/User';
 import { Link, useNavigate, useParams } from 'react-router-dom';
 
 function Friend() {
