+++ client/src/components/profile/MyFriends.jsx
@@ -2,13 +2,13 @@
 
 import { useState, useContext, useEffect } from 'react';
 import axios from 'axios';
-import '../App.css';
-import './css/clear.css'
-import './css/dark.css'
-import Header from './parts/Header';
-import FriendThumbnail from './parts/FriendThumbnail';
-import { DarkMode } from '../context/DarkMode';
-import { User } from '../context/User';
+import '../../App.css';
+import '../css/clear.css'
+import '../css/dark.css'
+import Header from '../parts/Header';
+import FriendThumbnail from '../parts/FriendThumbnail';
+import { DarkMode } from '../../context/DarkMode';
+import { User } from '../../context/User';
 import { Link, useNavigate, useParams } from 'react-router-dom';
 import { useTranslation } from "react-i18next"
